add raster VRT support and support for mapset names by interim results
